Unmark main as module
